lista_telefonica/lista_telefonica/requirements.txt
Django==3.2
asgiref==3.3.4
django-bootstrap-form==3.4
django-filter==2.4.0
django-phonenumber-field==5.0.0
django-widget-tweaks==1.4.8
phonenumbers==8.12.21
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1